# docs/swagger.yaml
basePath: /api/v1
definitions:
  helper.Meta:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
  helper.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/helper.Meta'
    type: object
...
  users.CheckEmail:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  users.LoginUserInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  users.RegisterUserInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
  users.User:
    properties:
      avatarFileName:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      passwordHash:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  users.UserFormatResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation of Ethereal Beauty.
  title: Ethereal Beauty Backend API Docs
  version: "1.0"
paths:
  /avatars:
    post:
      consumes:
      - multipart/form-data
      description: Upload account avatar
      parameters:
      - description: Request
        in: formData
        name: AvatarFileName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success to upload avatar
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/users.UserFormatResponse'
              type: object
        "400":
          description: Failed to upload avatar
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiAuthKey: []
      summary: Upload Avatar
      tags:
      - User
  /checkEmail:
    post:
      consumes:
      - application/json
      description: Check an email whether it's available or not
      parameters:
      - description: Request Body
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/users.CheckEmail'
      produces:
      - application/json
      responses:
        "200":
          description: Success to check email
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/users.UserFormatResponse'
              type: object
        "400":
          description: Failed to check email
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: Check Email Availability
      tags:
      - User
...
  /register:
    post:
      consumes:
      - application/json
      description: Register an account
      parameters:
      - description: Request Body
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/users.RegisterUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: Success to register an account
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/users.UserFormatResponse'
              type: object
        "400":
          description: Failed to register an account
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: Register User
      tags:
      - User
  /sessions:
    post:
      consumes:
      - application/json
      description: Login an account
      parameters:
      - description: Request Body
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/users.LoginUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: Success to login
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/users.UserFormatResponse'
              type: object
        "400":
          description: Failed to login
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: Login User
      tags:
      - User
...
  /users/fetch:
    post:
      consumes:
      - application/json
      description: Fetch User
      parameters:
      - description: Request Body
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/users.User'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch user
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/users.UserFormatResponse'
              type: object
        "400":
          description: Failed to fetch user
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiAuthKey: []
      summary: Fetch User
      tags:
      - User
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"